Add user-search bar and user-info
Added a search bar and the user information
(including avatar) in the first row of the header.
The second row is still in the works but three
buttons representing actions have been added
to the second column of the second row. A few
images have been added to a new folder "images",
including svgs