Make placeholder for project card
Ordered the items inside the first project card,
which will act as a template and placeholder for
other cards as well. Added an orange left
border to each card. Made the title bold
and description grey. Moved the options to
the bottom right by once again changing
place-items, as it was conflicting with the
header buttons selector. Changed text size for
both "Your projects" div and the text inside
the cards