Order layout of announcements
Used grid to order the announcement cards, added
a grey bottom border to each card that isn't
the last. Each card's title has a fixed size in
px, while the description fills up the rest of
the space available.
Added padding to the announcement container and
 made the announcement and
trending container take up more space in the page.