Make placeholder layout of trending
Made a first draft of the trending container
layout, although it doesn't differ too much
from the announcement container. First trend
card has been added along with an image which
will be replaced by the previous user-avatar
used in the header. User-name has been changed
to have a lighter weight and the card title
has been made grey, with the same color variant
as in the project cards (might have to
turn it into a variable)